Separated Genome and FitnessEvaluator into their own project as part of library
