+++ MultiStrand/MultiStrand.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30204.135
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MultiStrand", "MultiStrand\MultiStrand.csproj", "{EED1C598-3363-44DF-A508-F374BA24BE88}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MultiStrand", "MultiStrand\MultiStrand.csproj", "{EED1C598-3363-44DF-A508-F374BA24BE88}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MultiStrandLib", "MultiStrandLib\MultiStrandLib.csproj", "{D454EC89-1C5C-4E80-B4D7-9D37EC69CF32}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{EED1C598-3363-44DF-A508-F374BA24BE88}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EED1C598-3363-44DF-A508-F374BA24BE88}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EED1C598-3363-44DF-A508-F374BA24BE88}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D454EC89-1C5C-4E80-B4D7-9D37EC69CF32}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D454EC89-1C5C-4E80-B4D7-9D37EC69CF32}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D454EC89-1C5C-4E80-B4D7-9D37EC69CF32}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D454EC89-1C5C-4E80-B4D7-9D37EC69CF32}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
