+++ MultiStrand/MultiStrand/MultiStrand.csproj
@@ -5,4 +5,8 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\MultiStrandLib\MultiStrandLib.csproj" />
+  </ItemGroup>
+
 </Project>
